requirements.txt
# Core Dependencies
aiohttp>=3.9.0
asyncio-mqtt>=0.16.1
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9  # PostgreSQL support
redis>=5.0.1  # Caching & queue

# Network & Security
dnspython>=2.4.2
python-whois>=0.8.0
python-nmap>=0.7.1
scapy>=2.5.0
cryptography>=41.0.7
pyopenssl>=23.3.0

# HTTP & Web
requests>=2.31.0
httpx>=0.25.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
playwright>=1.40.0
selenium>=4.16.0

# Templates & Serialization
jinja2>=3.1.2
pyyaml>=6.0.1
toml>=0.10.2

# CLI & UI
rich>=13.7.0
click>=8.1.7
customtkinter>=5.2.1
textual>=0.46.0  # Modern TUI
prompt-toolkit>=3.0.43

# File & Data Processing
python-magic>=0.4.27
pandas>=2.1.4
openpyxl>=3.1.2

# Task Queue & Workers
celery>=5.3.4
kombu>=5.3.4

# API & Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2

# Security Tools Integration
python-libnmap>=0.7.3
shodan>=1.31.0
censys>=2.2.9

# Hashing & Cracking
passlib>=1.7.4
bcrypt>=4.1.2
hashlib-additional>=1.0.0

# Machine Learning (Phase 7)
scikit-learn>=1.3.2
numpy>=1.26.2

# PDF Generation
reportlab>=4.0.7
weasyprint>=60.1

# Testing (Development)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
colorama>=0.4.6
tabulate>=0.9.0
tqdm>=4.66.1
keyring>=24.3.0  # Secure credential storage
